Day31-40/Day40/requirements.txt
python-dotenv==1.0.1
Requests==2.32.3
twilio==9.1.1
